requirements.txt
# python==3.12.3
Django>=4.2
pycryptodome>=3.19.0
django-jazzmin
djangorestframework